C# 8 .NET Core 3 Preparations for some day replacing the expression pipeline with FastExpressionCompiler Nth rewrite of codegen Add facilities for detecting code unrolling - hopefully this helps with avoiding expression comparison I'm not really sure what else is in this commit

+++ DBClientFiles.NET/Parsing/Shared/Segments/Handlers/Implementations/ExtendedFieldInfoHandler.cs
@@ -93,7 +93,7 @@ namespace DBClientFiles.NET.Parsing.Shared.Segments.Handlers.Implementations
             // fieldInfoSize == 0 means the field is outside the record (it's either a relationship column, or our index)
             // We don't bother forcing the calculation if arity was already specified in the field info.
             if (currentField.CompressionData.Type != MemberCompressionType.BitpackedPalletArrayData && fieldInfoSize != 0)
-                currentField.Cardinality = (int) (currentField.Size / fieldInfoSize);
+                currentField.Cardinality = currentField.Size / fieldInfoSize;
 
             return currentField;
         }
